Refactor health monitoring core into modular components
