+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -169,6 +169,13 @@
 - **Completion Date**: 2025-08-24
 - **Summary**: Refactored from 722 to 55 lines by extracting manager, AI processor, coordinator, and config modules. New orchestrator coordinates components and maintains functionality.
 
+### MODERATE-011: Health Monitoring Core Split ✅
+- **File**: `src/core/health/monitoring/core.py`
+- **Status**: Completed
+- **Assigned To**: Victor Dixon
+- **Completion Date**: 2025-08-24
+- **Summary**: Split monolithic core into dedicated core, metrics, alerts, and config modules. New orchestrator imports these components to coordinate health monitoring.
+
 ## 📋 AVAILABLE CONTRACTS FOR CLAIMING
 
 > **Note:** Remaining contract descriptions include current line counts for context only. There is no strict LOC target—deliver clean, production-ready, tested modules that honor SRP and SOLID principles.
